refactor(redis_ops): extract batch module
Extract batch optimization functions and KeyMetadata struct into batch.rs:
- KeyMetadata: Struct containing key name, type, and TTL
- fetch_key_metadata: Retrieves TYPE and TTL for batch of keys
- partition_keys_by_type: Splits keys into strings vs others
- process_string_keys_batch: Optimized batch GET for strings
- process_non_string_key: Routes other types to appropriate handler
- generate_redis_commands_batch_optimized: Main optimization logic

All 4 partition tests moved with the functions.
